+++ src/components/GameCard/GameCard.js
@@ -1,13 +1,61 @@
 import { useEffect } from "react";
 import "./GameCard.scss";
 
-const GameCard = ({ checkIsVisible, card, getActiveCards, onCardClick }) => {
-  const {id, isHidden } = card;
+const GameCard = ({ isHidden, cardsData, setCardsData, id, card }) => {
+  const onCardClick = () => {
+    setCardsData(
+      cardsData.map((item) =>
+        item.id === id ? { ...item, isHidden: !isHidden } : item
+      )
+    );
+  };
+
+  const getActiveCards = () => {
+    return cardsData.filter(
+      (card) => card.isHidden === false && card.guessed === false
+    );
+  };
+
+  const checkIsVisible = (visibleCards) => {
+    if (visibleCards.length === 2) {
+      let areIdenticalValues = true;
+
+      visibleCards.forEach((visibleCard) => {
+        if (visibleCard.value !== card.value) {
+          areIdenticalValues = false;
+        }
+      });
+
+      if (areIdenticalValues) {
+        const newCards = cardsData.map((card) => {
+          if (card.isHidden === false) {
+            return {
+              ...card,
+              guessed: true,
+            };
+          }
+          return card;
+        });
+        setCardsData(newCards);
+      }
+
+      if (!areIdenticalValues) {
+        const newCards = cardsData.map((card) => {
+          if (card.isHidden === false && card.guessed === false) {
+            return { ...card, isHidden: true };
+          }
+
+          return card;
+        });
+
+        setTimeout(() => setCardsData(newCards), 500);
+      }
+    }
+  };
 
   const updateCards = () => {
     const activeCards = getActiveCards();
-
-    checkIsVisible(activeCards, card);
+    checkIsVisible(activeCards);
   };
 
   const getCardStyles = () => {
@@ -27,7 +75,7 @@ const GameCard = ({ checkIsVisible, card, getActiveCards, onCardClick }) => {
   return (
     <div
       className={getCardStyles()}
-      onClick={!card.guessed ? () => onCardClick(id, isHidden) : () => console.log("guessed")}
+      onClick={!card.guessed ? onCardClick : () => console.log("guessed")}
     >
       {!isHidden && card.value}
     </div>
